Multi-type tuner workaround.
For boxes with multi-type tuners that support satellite, terrestrial and cable. Due to E2 limitations it is only possible to configure one DVB tuner type at once, but these tuners are capable of tuning satellite and terrestrial, or satellite and cable at the same time without reconfiguation. This workaround allows ABM to recognise the tuning capabilities of the said tuners. Not a perfect solution but until E2 gets the total redisign of the UI in order to support multiple delivery systems this is the best that is possible at this time. All conventional and DVB-T2/C hybrid tuners will not be affected by this workaround.