Process multiple NIT-other tables
When multiple NIT-other tables (for multiple network IDs) are being
published in the transport stream, try to process all of them rather
than the previous behaviour of processing only the first one seen.

Every NIT-other table received will be processed until one of the
following conditions is true:

- A NIT-other section is received with a network ID and version number
  matching one we've already processed, and all NIT-other tables we've
  seen are complete (we've received and processed all the sections)
- TIMEOUT_SEC has elapsed

Not having a definitive way to know how many different NIT-other tables
are being published, this assumes that the NIT-other tables are
transmitted as a repeating loop with network IDs in the same order each
time. This matches the observed behaviour of every real-world NIT-other
stream I've examined.