Remove name added for 53595
The name I gave 53595 "The Craft Store HD" does not match the name it got when it was given a real name.
Though as far as I can see, my name actually shows what it (it seems to be an HD version of "The Craft Store").
But, at least until I'm sure what they're up to, I suggest we remove the name I gave it for now.